requirements.txt
discord.py
requests
aiohttp
pendulum
flask
jicson